CI Improvements

+++ src/map/lua/luautils.cpp
@@ -4277,7 +4277,7 @@ namespace luautils
     {
         TPZ_DEBUG_BREAK_IF(lua_isnil(L, -1) || !lua_isnumber(L, -1));
 
-        uint32 id = lua_tointeger(L, 1);
+        uint32 id = static_cast<uint32>(lua_tointeger(L, 1));
         CItem* PItem = itemutils::GetItemPointer(id);
         if (PItem)
         {
